consistency in namespace conventions of test classes
- namespace now always starts with Tests\* instead of the vendor namespace (i.e. Shopsys\*)
- namespace always contains the rest of root namespace
  - eg. package with source root Shopsys\ProductFeed\GoogleBundle has test namespace Tests\ProductFeed\GoogleBundle
- namespace always contains the test type: Unit / Database / Smoke / Acceptance / Performance
  - packages contain only unit tests at the moment
