+++ packages/http-smoke-testing/composer.json
@@ -15,6 +15,11 @@
             "Shopsys\\HttpSmokeTesting\\": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\HttpSmokeTesting\\": "tests/"
+        }
+    },
     "require": {
         "php": "^5.6|^7.0",
         "symfony/framework-bundle": "^2.5|^3.0|^4.0",
