+++ packages/product-feed-zbozi/composer.json
@@ -17,7 +17,7 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Tests\\": "tests/"
+            "Tests\\ProductFeed\\ZboziBundle\\": "tests/"
         }
     },
     "require": {
